Fix memory and debounce download updates

+++ crates/carbon_app/src/api/instance.rs
@@ -126,7 +126,7 @@ pub(super) fn mount() -> impl RouterBuilderLike<App> {
                     details.use_loaded_icon.into(),
                     details.notes.into(),
                     details.memory.into_option()
-                        .map(|m| m.map(|(xms, xmx)| (xms as u64, xmx as u64))),
+                        .map(|m| m.map(|(xms, xmx)| (xms, xmx))),
                 )
                 .await
         }
@@ -400,7 +400,7 @@ struct UpdateInstance {
     name: Update<String>,
     use_loaded_icon: Update<bool>,
     notes: Update<String>,
-    memory: Update<Option<(u32, u32)>>,
+    memory: Update<Option<(u16, u16)>>,
 }
 
 #[derive(Type, Deserialize)]
